requirements.txt
Flask>=2.0
Flask-CORS>=3.0
google-cloud-translate>=2.0.1
